Move input/output functions to new file
