+++ src/include/coati/align_msa.hpp
@@ -25,6 +25,7 @@
 
 #include "align_pair.hpp"
 #include "insertions.hpp"
+#include "io.hpp"
 #include "semiring.hpp"
 #include "tree.hpp"
 #include "utils.hpp"
