+++ src/lib/align_msa.cc
@@ -97,7 +97,7 @@ bool ref_indel_alignment(coati::alignment_t& input) {
     }
 
     // write alignment
-    coati::utils::write_output(aln.data);
+    coati::io::write_output(aln.data);
     return true;
 }
